devtools/client/shared/components/test/mochitest/test_stack-trace.html

<!DOCTYPE HTML>
<html>
<!--
Test the rendering of a stack trace
-->
<head>
  <meta charset="utf-8">
  <title>StackTrace component test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<script src="head.js"></script>
<script>
window.onload = function() {
  let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
  let React = browserRequire("devtools/client/shared/vendor/react");
  let StackTrace = React.createFactory(browserRequire("devtools/client/shared/components/stack-trace"));
  ok(StackTrace, "Got the StackTrace factory");

  add_task(function* () {
    let stacktrace = [
      {
        filename: "http://myfile.com/mahscripts.js",
        lineNumber: 55,
        columnNumber: 10
      },
      {
        asyncCause: "because",
        functionName: "loadFunc",
        filename: "http://myfile.com/loader.js -> http://myfile.com/loadee.js",
        lineNumber: 10
      }
    ];

    let props = {
      stacktrace,
      onViewSourceInDebugger: () => {}
    };

    let trace = ReactDOM.render(StackTrace(props), window.document.body);
    yield forceRender(trace);

    let traceEl = trace.getDOMNode();
    ok(traceEl, "Rendered StackTrace has an element");

    let frameEls = traceEl.childNodes;
    ok(frameEls, "Rendered StackTrace has frames");
    is(frameEls.length, 3, "StackTrace has 3 frames");

    // Check the top frame, function name should be anonymous
    checkFrameString({
      el: frameEls[0],
      functionName: "<anonymous>",
      source: "http://myfile.com/mahscripts.js",
      file: "http://myfile.com/mahscripts.js",
      line: 55,
      column: 10,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/mahscripts.js:55:10",
    });

    // Check the async cause node
    is(frameEls[1].className, "frame-link-async-cause", "Async cause has the right class");
    is(frameEls[1].textContent, "(Async: because)", "Async cause has the right label");

    // Check the third frame, the source should be parsed into a valid source URL
    checkFrameString({
      el: frameEls[2],
      functionName: "loadFunc",
      source: "http://myfile.com/loadee.js",
      file: "http://myfile.com/loadee.js",
      line: 10,
      column: null,
      shouldLink: true,
      tooltip: "View source in Debugger → http://myfile.com/loadee.js:10",
    });
  });
}
</script>
</body>
</html>
